+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* WP Rocket: exclude inline JS from minification
+
 ### 1.2.2: September 14th, 2021
 * Fix minified public javascript causing javascript errors